requirements.txt
Pillow
av
deepface
numpy
opencv-python
pandas
psycopg2-binary
requests
streamlit
streamlit-webrtc
tensorflow
transformers